+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/error/GraphQLReferencedResourceNotFoundErrorBuilder.java
@@ -15,7 +15,7 @@ import io.vrap.rmf.base.client.utils.Generated;
  * <div class=code-example>
  * <pre><code class='java'>
  *     GraphQLReferencedResourceNotFoundError graphQLReferencedResourceNotFoundError = GraphQLReferencedResourceNotFoundError.builder()
- *             .typeId(ReferenceTypeId.ASSOCIATE_ROLE)
+ *             .typeId(ReferenceTypeId.APPROVAL_FLOW)
  *             .build()
  * </code></pre>
  * </div>